requirements.txt
Click
PyYAML
boto3
botocore
requests
